Add a decoder for any Jelly streams
The new decoder auto-detects the type of the stream and then acts accordingly. This will be needed for integrations with Jena/RDF4J where we'll need to guess the type of the stream... or create a ton of separate formats (ugh).

this and some refactors
