+++ core/src/test/scala/eu/ostrzyciel/jelly/core/ProtoDecoderSpec.scala
@@ -8,7 +8,7 @@ import org.scalatest.matchers.should.Matchers
 import org.scalatest.wordspec.AnyWordSpec
 
 class ProtoDecoderSpec extends AnyWordSpec, Matchers:
-  import ProtoDecoder.*
+  import ProtoDecoderImpl.*
   import ProtoTestCases.*
 
   // Test body
