+++ jena/src/main/scala/eu/ostrzyciel/jelly/convert/jena/riot/JellyLanguage.scala
@@ -1,9 +1,9 @@
-package pl.ostrzyciel.jelly.convert.jena.riot
+package eu.ostrzyciel.jelly.convert.jena.riot
 
 import org.apache.jena.riot.{LangBuilder, RDFLanguages, RDFParserRegistry}
 
 object JellyLanguage:
-  val contentTypeJelly = ""
+  val contentTypeJelly = "application/x-jelly-rdf"
   val strLangJelly = "Jelly"
 
   val JELLY = LangBuilder.create(strLangJelly, contentTypeJelly)
@@ -15,6 +15,3 @@ object JellyLanguage:
   RDFLanguages.register(JELLY)
   RDFParserRegistry.registerLangTriples(JELLY, JellyReaderFactory)
   RDFParserRegistry.registerLangQuads(JELLY, JellyReaderFactory)
-
-
-
